Updates helpers because Balance model amount property is a Double type now, rather than Int .

+++ SolanaSwiftSDK/SolanaPackage/SolanaPackageTests/Balance API/BalanceViewModelMapperTests.swift
@@ -26,7 +26,7 @@ class BalanceViewModelMapperTests: XCTestCase {
         XCTAssertEqual(viewModel, result.viewModel)
     }
     
-    private func makeModelandViewModel(amount: Int) -> (model: Balance, viewModel: BalanceViewModel) {
+    private func makeModelandViewModel(amount: Double) -> (model: Balance, viewModel: BalanceViewModel) {
         let model = Balance(amount: amount)
         let viewModel = BalanceViewModel(amount: "\(model.amount)")
